pyproject.toml
[project]
name = "s2-assignment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.0.0",
    "pillow>=10.0.0",
    "numpy>=1.24.0",
    "werkzeug>=3.0.0"
]
